premium_store_backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
Django==6.0
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-filter==25.2
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.95.0
exceptiongroup==1.3.1
idna==3.11
kavenegar==1.1.2
kombu==5.6.1
oauthlib==3.3.1
packaging==25.0
phonenumbers==9.0.19
pillow==12.0.0
prompt_toolkit==3.0.52
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.4
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.1
vine==5.1.0
wcwidth==0.2.14
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
